Create tests for using cache and coalescing mutations
